test: centralize Anthropic container recovery coverage
